+++ kernel/fs/squashfs/symlink.c
@@ -49,7 +49,7 @@ static int squashfs_symlink_readpage(struct file *file, struct page *page)
 	struct super_block *sb = inode->i_sb;
 	struct squashfs_sb_info *msblk = sb->s_fs_info;
 	int index = page->index << PAGE_CACHE_SHIFT;
-	long long block = squashfs_i(inode)->start;
+	u64 block = squashfs_i(inode)->start;
 	int offset = squashfs_i(inode)->offset;
 	int length = min_t(int, i_size_read(inode) - index, PAGE_CACHE_SIZE);
 	int bytes, copied;
